TYP: minor fixes related to ``errstate``
